requirements.txt
# Core
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0

# Database
sqlite3 (built-in)

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# File handling
pathlib (built-in)

# Logging
loguru>=0.7.0